Move to big decimal

+++ golem-rib/src/parser/block.rs
@@ -22,6 +22,7 @@ where
 
 #[cfg(test)]
 mod tests {
+    use bigdecimal::BigDecimal;
     use test_r::test;
 
     use super::*;
@@ -49,8 +50,8 @@ mod tests {
         let expr = block().easy_parse(input).unwrap().0;
 
         let expected = Expr::expr_block(vec![
-            Expr::let_binding("x", Expr::untyped_number(1f64)),
-            Expr::let_binding("y", Expr::untyped_number(2f64)),
+            Expr::let_binding("x", Expr::untyped_number(BigDecimal::from(1))),
+            Expr::let_binding("y", Expr::untyped_number(BigDecimal::from(2))),
             Expr::plus(Expr::identifier("x"), Expr::identifier("y")),
         ]);
         assert_eq!(expr, expected);
