+++ golem-rib/src/type_inference/variable_binding_pattern_match.rs
@@ -258,6 +258,7 @@ mod pattern_match_bindings {
     }
 
     mod expectations {
+        use bigdecimal::BigDecimal;
         use crate::{ArmPattern, Expr, InferredType, MatchArm, MatchIdentifier, VariableId};
 
         pub(crate) fn expected_match(index: usize) -> Expr {
@@ -291,7 +292,7 @@ mod pattern_match_bindings {
                     },
                     MatchArm {
                         arm_pattern: ArmPattern::constructor("none", vec![]),
-                        arm_resolution_expr: Box::new(Expr::untyped_number(0f64)),
+                        arm_resolution_expr: Box::new(Expr::untyped_number(BigDecimal::from(0))),
                     },
                 ],
                 InferredType::Unknown,
@@ -299,7 +300,7 @@ mod pattern_match_bindings {
         }
 
         pub(crate) fn expected_match_with_let_binding(index: usize) -> Expr {
-            let let_binding = Expr::let_binding("x", Expr::untyped_number(1f64));
+            let let_binding = Expr::let_binding("x", Expr::untyped_number(BigDecimal::from(1)));
             let identifier_expr =
                 Expr::Identifier(VariableId::Global("x".to_string()), InferredType::Unknown);
             let block = Expr::ExprBlock(vec![let_binding, identifier_expr], InferredType::Unknown);
@@ -322,7 +323,7 @@ mod pattern_match_bindings {
                     },
                     MatchArm {
                         arm_pattern: ArmPattern::constructor("none", vec![]),
-                        arm_resolution_expr: Box::new(Expr::untyped_number(0f64)),
+                        arm_resolution_expr: Box::new(Expr::untyped_number(BigDecimal::from(0))),
                     },
                 ],
                 InferredType::Unknown,
@@ -388,7 +389,7 @@ mod pattern_match_bindings {
                                 },
                                 MatchArm {
                                     arm_pattern: ArmPattern::constructor("none", vec![]),
-                                    arm_resolution_expr: Box::new(Expr::untyped_number(0f64)),
+                                    arm_resolution_expr: Box::new(Expr::untyped_number(BigDecimal::from(0))),
                                 },
                             ],
                             InferredType::Unknown,
@@ -405,7 +406,7 @@ mod pattern_match_bindings {
                                 InferredType::Unknown,
                             ))],
                         ),
-                        arm_resolution_expr: Box::new(Expr::untyped_number(0f64)),
+                        arm_resolution_expr: Box::new(Expr::untyped_number(BigDecimal::from(0))),
                     },
                 ],
                 InferredType::Unknown,
